Extracted pushToRemote in scaffolder-backend
