+++ .changeset/silent-suns-bathe.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-scaffolder-backend': patch
+---
+
+Extracted pushToRemote function for reuse between publishers
