+++ plugins/scaffolder-backend/src/scaffolder/stages/publish/azure.test.ts
@@ -17,34 +17,18 @@
 jest.mock('nodegit');
 jest.mock('azure-devops-node-api/GitApi');
 jest.mock('azure-devops-node-api/interfaces/GitInterfaces');
+jest.mock('./helpers', () => ({
+  pushToRemoteUserPass: jest.fn(),
+}));
 
 import { AzurePublisher } from './azure';
 import { GitApi } from 'azure-devops-node-api/GitApi';
-import * as NodeGit from 'nodegit';
+import { pushToRemoteUserPass } from './helpers';
 
 const { mockGitApi } = require('azure-devops-node-api/GitApi') as {
   mockGitApi: {
     createRepository: jest.MockedFunction<GitApi['createRepository']>;
   };
-};
-
-const {
-  Repository,
-  mockRepo,
-  mockIndex,
-  Signature,
-  Remote,
-  mockRemote,
-  Cred,
-} = require('nodegit') as {
-  Repository: jest.Mocked<{ init: any }>;
-  Signature: jest.Mocked<{ now: any }>;
-  Cred: jest.Mocked<{ userpassPlaintextNew: any }>;
-  Remote: jest.Mocked<{ create: any }>;
-
-  mockIndex: jest.Mocked<NodeGit.Index>;
-  mockRepo: jest.Mocked<NodeGit.Repository>;
-  mockRemote: jest.Mocked<NodeGit.Remote>;
 };
 
 describe('Azure Publisher', () => {
@@ -60,7 +44,7 @@ describe('Azure Publisher', () => {
         remoteUrl: 'mockclone',
       } as { remoteUrl: string });
 
-      await publisher.publish({
+      const result = await publisher.publish({
         values: {
           storePath: 'project/repo',
           owner: 'bob',
@@ -68,109 +52,16 @@ describe('Azure Publisher', () => {
         directory: '/tmp/test',
       });
 
+      expect(result).toEqual({ remoteUrl: 'mockclone' });
       expect(mockGitApi.createRepository).toHaveBeenCalledWith(
         {
           name: 'repo',
         },
         'project',
       );
-    });
-  });
-
-  describe('publish: createGitDirectory', () => {
-    const values = {
-      isOrg: true,
-      storePath: 'blam/test',
-      owner: 'lols',
-    };
-
-    const mockDir = '/tmp/test/dir';
-
-    mockGitApi.createRepository.mockResolvedValue({
-      remoteUrl: 'mockclone',
-    } as { remoteUrl: string });
-
-    it('should call init on the repo with the directory', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(Repository.init).toHaveBeenCalledWith(mockDir, 0);
-    });
-
-    it('should call refresh index on the index and write the new files', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(mockRepo.refreshIndex).toHaveBeenCalled();
-    });
-
-    it('should call add all files and write', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(mockIndex.addAll).toHaveBeenCalled();
-      expect(mockIndex.write).toHaveBeenCalled();
-      expect(mockIndex.writeTree).toHaveBeenCalled();
-    });
-
-    it('should create a commit with on head with the right name and commiter', async () => {
-      const mockSignature = { mockSignature: 'bloblly' };
-      Signature.now.mockReturnValue(mockSignature);
-
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(Signature.now).toHaveBeenCalledTimes(2);
-      expect(Signature.now).toHaveBeenCalledWith(
-        'Scaffolder',
-        'scaffolder@backstage.io',
-      );
-
-      expect(mockRepo.createCommit).toHaveBeenCalledWith(
-        'HEAD',
-        mockSignature,
-        mockSignature,
-        'initial commit',
-        'mockoid',
-        [],
-      );
-    });
-
-    it('creates a remote with the repo and remote', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(Remote.create).toHaveBeenCalledWith(
-        mockRepo,
-        'origin',
+      expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+        '/tmp/test',
         'mockclone',
-      );
-    });
-
-    it('shoud push to the remote repo', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      const [remotes, { callbacks }] = mockRemote.push.mock
-        .calls[0] as NodeGit.PushOptions[];
-
-      expect(remotes).toEqual(['refs/heads/master:refs/heads/master']);
-
-      callbacks?.credentials?.();
-
-      expect(Cred.userpassPlaintextNew).toHaveBeenCalledWith(
         'notempty',
         'fake-token',
       );
