+++ plugins/scaffolder-backend/src/scaffolder/stages/publish/github.test.ts
@@ -16,15 +16,18 @@
 
 jest.mock('@octokit/rest');
 jest.mock('nodegit');
+jest.mock('./helpers', () => ({
+  pushToRemoteUserPass: jest.fn(),
+}));
 
 import { Octokit } from '@octokit/rest';
-import * as NodeGit from 'nodegit';
 import {
   OctokitResponse,
   ReposCreateInOrgResponseData,
   UsersGetByUsernameResponseData,
 } from '@octokit/types';
 import { GithubPublisher } from './github';
+import { pushToRemoteUserPass } from './helpers';
 
 const { mockGithubClient } = require('@octokit/rest') as {
   mockGithubClient: {
@@ -32,25 +35,6 @@ const { mockGithubClient } = require('@octokit/rest') as {
     users: jest.Mocked<Octokit['users']>;
     teams: jest.Mocked<Octokit['teams']>;
   };
-};
-
-const {
-  Repository,
-  mockRepo,
-  mockIndex,
-  Signature,
-  Remote,
-  mockRemote,
-  Cred,
-} = require('nodegit') as {
-  Repository: jest.Mocked<{ init: any }>;
-  Signature: jest.Mocked<{ now: any }>;
-  Cred: jest.Mocked<{ userpassPlaintextNew: any }>;
-  Remote: jest.Mocked<{ create: any }>;
-
-  mockIndex: jest.Mocked<NodeGit.Index>;
-  mockRepo: jest.Mocked<NodeGit.Repository>;
-  mockRemote: jest.Mocked<NodeGit.Remote>;
 };
 
 describe('GitHub Publisher', () => {
@@ -72,8 +56,13 @@ describe('GitHub Publisher', () => {
             clone_url: 'mockclone',
           },
         } as OctokitResponse<ReposCreateInOrgResponseData>);
+        mockGithubClient.users.getByUsername.mockResolvedValue({
+          data: {
+            type: 'Organization',
+          },
+        } as OctokitResponse<UsersGetByUsernameResponseData>);
 
-        await publisher.publish({
+        const result = await publisher.publish({
           values: {
             storePath: 'blam/test',
             owner: 'bob',
@@ -82,6 +71,7 @@ describe('GitHub Publisher', () => {
           directory: '/tmp/test',
         });
 
+        expect(result).toEqual({ remoteUrl: 'mockclone' });
         expect(mockGithubClient.repos.createInOrg).toHaveBeenCalledWith({
           org: 'blam',
           name: 'test',
@@ -97,6 +87,12 @@ describe('GitHub Publisher', () => {
           repo: 'test',
           permission: 'admin',
         });
+        expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+          '/tmp/test',
+          'mockclone',
+          'abc',
+          'x-oauth-basic',
+        );
       });
 
       it('should use octokit to create a repo in the authed user if the organisation property is not set', async () => {
@@ -111,7 +107,7 @@ describe('GitHub Publisher', () => {
           },
         } as OctokitResponse<UsersGetByUsernameResponseData>);
 
-        await publisher.publish({
+        const result = await publisher.publish({
           values: {
             storePath: 'blam/test',
             owner: 'bob',
@@ -120,6 +116,7 @@ describe('GitHub Publisher', () => {
           directory: '/tmp/test',
         });
 
+        expect(result).toEqual({ remoteUrl: 'mockclone' });
         expect(
           mockGithubClient.repos.createForAuthenticatedUser,
         ).toHaveBeenCalledWith({
@@ -127,6 +124,12 @@ describe('GitHub Publisher', () => {
           private: false,
         });
         expect(mockGithubClient.repos.addCollaborator).not.toHaveBeenCalled();
+        expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+          '/tmp/test',
+          'mockclone',
+          'abc',
+          'x-oauth-basic',
+        );
       });
     });
 
@@ -142,7 +145,7 @@ describe('GitHub Publisher', () => {
         },
       } as OctokitResponse<UsersGetByUsernameResponseData>);
 
-      await publisher.publish({
+      const result = await publisher.publish({
         values: {
           storePath: 'blam/test',
           owner: 'bob',
@@ -152,6 +155,7 @@ describe('GitHub Publisher', () => {
         directory: '/tmp/test',
       });
 
+      expect(result).toEqual({ remoteUrl: 'mockclone' });
       expect(
         mockGithubClient.repos.createForAuthenticatedUser,
       ).toHaveBeenCalledWith({
@@ -165,113 +169,12 @@ describe('GitHub Publisher', () => {
         username: 'bob',
         permission: 'admin',
       });
-    });
-
-    describe('publish: createGitDirectory', () => {
-      const values = {
-        storePath: 'blam/test',
-        owner: 'lols',
-        access: 'lols',
-      };
-
-      const mockDir = '/tmp/test/dir';
-
-      mockGithubClient.repos.createInOrg.mockResolvedValue({
-        data: {
-          clone_url: 'mockclone',
-        },
-      } as OctokitResponse<ReposCreateInOrgResponseData>);
-      mockGithubClient.users.getByUsername.mockResolvedValue({
-        data: {
-          type: 'Organization',
-        },
-      } as OctokitResponse<UsersGetByUsernameResponseData>);
-
-      it('should call init on the repo with the directory', async () => {
-        await publisher.publish({
-          values,
-          directory: mockDir,
-        });
-
-        expect(Repository.init).toHaveBeenCalledWith(mockDir, 0);
-      });
-
-      it('should call refresh index on the index and write the new files', async () => {
-        await publisher.publish({
-          values,
-          directory: mockDir,
-        });
-
-        expect(mockRepo.refreshIndex).toHaveBeenCalled();
-      });
-
-      it('should call add all files and write', async () => {
-        await publisher.publish({
-          values,
-          directory: mockDir,
-        });
-
-        expect(mockIndex.addAll).toHaveBeenCalled();
-        expect(mockIndex.write).toHaveBeenCalled();
-        expect(mockIndex.writeTree).toHaveBeenCalled();
-      });
-
-      it('should create a commit with on head with the right name and commiter', async () => {
-        const mockSignature = { mockSignature: 'bloblly' };
-        Signature.now.mockReturnValue(mockSignature);
-
-        await publisher.publish({
-          values,
-          directory: mockDir,
-        });
-
-        expect(Signature.now).toHaveBeenCalledTimes(2);
-        expect(Signature.now).toHaveBeenCalledWith(
-          'Scaffolder',
-          'scaffolder@backstage.io',
-        );
-
-        expect(mockRepo.createCommit).toHaveBeenCalledWith(
-          'HEAD',
-          mockSignature,
-          mockSignature,
-          'initial commit',
-          'mockoid',
-          [],
-        );
-      });
-
-      it('creates a remote with the repo and remote', async () => {
-        await publisher.publish({
-          values,
-          directory: mockDir,
-        });
-
-        expect(Remote.create).toHaveBeenCalledWith(
-          mockRepo,
-          'origin',
-          'mockclone',
-        );
-      });
-
-      it('shoud push to the remote repo', async () => {
-        await publisher.publish({
-          values,
-          directory: mockDir,
-        });
-
-        const [remotes, { callbacks }] = mockRemote.push.mock
-          .calls[0] as NodeGit.PushOptions[];
-
-        expect(remotes).toEqual(['refs/heads/master:refs/heads/master']);
-
-        callbacks?.credentials?.();
-
-        expect(Cred.userpassPlaintextNew).toHaveBeenCalledWith(
-          'abc',
-          'x-oauth-basic',
-        );
-      });
+      expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+        '/tmp/test',
+        'mockclone',
+        'abc',
+        'x-oauth-basic',
+      );
     });
   });
 
@@ -294,7 +197,7 @@ describe('GitHub Publisher', () => {
         },
       } as OctokitResponse<UsersGetByUsernameResponseData>);
 
-      await publisher.publish({
+      const result = await publisher.publish({
         values: {
           isOrg: true,
           storePath: 'blam/test',
@@ -303,12 +206,19 @@ describe('GitHub Publisher', () => {
         directory: '/tmp/test',
       });
 
+      expect(result).toEqual({ remoteUrl: 'mockclone' });
       expect(mockGithubClient.repos.createInOrg).toHaveBeenCalledWith({
         org: 'blam',
         name: 'test',
         private: true,
         visibility: 'internal',
       });
+      expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+        '/tmp/test',
+        'mockclone',
+        'abc',
+        'x-oauth-basic',
+      );
     });
   });
 
@@ -331,7 +241,7 @@ describe('GitHub Publisher', () => {
         },
       } as OctokitResponse<UsersGetByUsernameResponseData>);
 
-      await publisher.publish({
+      const result = await publisher.publish({
         values: {
           storePath: 'blam/test',
           owner: 'bob',
@@ -339,12 +249,19 @@ describe('GitHub Publisher', () => {
         directory: '/tmp/test',
       });
 
+      expect(result).toEqual({ remoteUrl: 'mockclone' });
       expect(
         mockGithubClient.repos.createForAuthenticatedUser,
       ).toHaveBeenCalledWith({
         name: 'test',
         private: true,
       });
+      expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+        '/tmp/test',
+        'mockclone',
+        'abc',
+        'x-oauth-basic',
+      );
     });
   });
 });
