+++ plugins/scaffolder-backend/src/scaffolder/stages/publish/github.ts
@@ -16,10 +16,9 @@
 
 import { PublisherBase } from './types';
 import { Octokit } from '@octokit/rest';
-
+import { pushToRemoteUserPass } from './helpers';
 import { JsonValue } from '@backstage/config';
 import { RequiredTemplateValues } from '../templater';
-import { Repository, Remote, Signature, Cred } from 'nodegit';
 
 export type RepoVisibilityOptions = 'private' | 'internal' | 'public';
 
@@ -52,7 +51,12 @@ export class GithubPublisher implements PublisherBase {
     directory: string;
   }): Promise<{ remoteUrl: string }> {
     const remoteUrl = await this.createRemote(values);
-    await this.pushToRemote(directory, remoteUrl);
+    await pushToRemoteUserPass(
+      directory,
+      remoteUrl,
+      this.token,
+      'x-oauth-basic',
+    );
 
     return { remoteUrl };
   }
@@ -103,30 +107,5 @@ export class GithubPublisher implements PublisherBase {
     }
 
     return data?.clone_url;
-  }
-
-  private async pushToRemote(directory: string, remote: string): Promise<void> {
-    const repo = await Repository.init(directory, 0);
-    const index = await repo.refreshIndex();
-    await index.addAll();
-    await index.write();
-    const oid = await index.writeTree();
-    await repo.createCommit(
-      'HEAD',
-      Signature.now('Scaffolder', 'scaffolder@backstage.io'),
-      Signature.now('Scaffolder', 'scaffolder@backstage.io'),
-      'initial commit',
-      oid,
-      [],
-    );
-
-    const remoteRepo = await Remote.create(repo, 'origin', remote);
-    await remoteRepo.push(['refs/heads/master:refs/heads/master'], {
-      callbacks: {
-        credentials: () => {
-          return Cred.userpassPlaintextNew(this.token, 'x-oauth-basic');
-        },
-      },
-    });
   }
 }
