+++ plugins/scaffolder-backend/src/scaffolder/stages/publish/gitlab.test.ts
@@ -16,11 +16,14 @@
 
 jest.mock('nodegit');
 jest.mock('@gitbeaker/node');
+jest.mock('./helpers', () => ({
+  pushToRemoteUserPass: jest.fn(),
+}));
 
 import { GitlabPublisher } from './gitlab';
 import { Gitlab as GitlabAPI } from '@gitbeaker/core';
 import { Gitlab } from '@gitbeaker/node';
-import * as NodeGit from 'nodegit';
+import { pushToRemoteUserPass } from './helpers';
 
 const { mockGitlabClient } = require('@gitbeaker/node') as {
   mockGitlabClient: {
@@ -28,25 +31,6 @@ const { mockGitlabClient } = require('@gitbeaker/node') as {
     Projects: jest.Mocked<GitlabAPI['Projects']>;
     Users: jest.Mocked<GitlabAPI['Users']>;
   };
-};
-
-const {
-  Repository,
-  mockRepo,
-  mockIndex,
-  Signature,
-  Remote,
-  mockRemote,
-  Cred,
-} = require('nodegit') as {
-  Repository: jest.Mocked<{ init: any }>;
-  Signature: jest.Mocked<{ now: any }>;
-  Cred: jest.Mocked<{ userpassPlaintextNew: any }>;
-  Remote: jest.Mocked<{ create: any }>;
-
-  mockIndex: jest.Mocked<NodeGit.Index>;
-  mockRepo: jest.Mocked<NodeGit.Repository>;
-  mockRemote: jest.Mocked<NodeGit.Remote>;
 };
 
 describe('GitLab Publisher', () => {
@@ -61,8 +45,11 @@ describe('GitLab Publisher', () => {
       mockGitlabClient.Namespaces.show.mockResolvedValue({
         id: 42,
       } as { id: number });
+      mockGitlabClient.Projects.create.mockResolvedValue({
+        http_url_to_repo: 'mockclone',
+      } as { http_url_to_repo: string });
 
-      await publisher.publish({
+      const result = await publisher.publish({
         values: {
           isOrg: true,
           storePath: 'blam/test',
@@ -71,10 +58,17 @@ describe('GitLab Publisher', () => {
         directory: '/tmp/test',
       });
 
+      expect(result).toEqual({ remoteUrl: 'mockclone' });
       expect(mockGitlabClient.Projects.create).toHaveBeenCalledWith({
         namespace_id: 42,
         name: 'test',
       });
+      expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+        '/tmp/test',
+        'mockclone',
+        'oauth2',
+        'fake-token',
+      );
     });
 
     it('should use gitbeaker to create a repo in the authed user if the namespace property is not set', async () => {
@@ -86,7 +80,7 @@ describe('GitLab Publisher', () => {
         http_url_to_repo: 'mockclone',
       } as { http_url_to_repo: string });
 
-      await publisher.publish({
+      const result = await publisher.publish({
         values: {
           storePath: 'blam/test',
           owner: 'bob',
@@ -94,109 +88,15 @@ describe('GitLab Publisher', () => {
         directory: '/tmp/test',
       });
 
+      expect(result).toEqual({ remoteUrl: 'mockclone' });
       expect(mockGitlabClient.Users.current).toHaveBeenCalled();
-
       expect(mockGitlabClient.Projects.create).toHaveBeenCalledWith({
         namespace_id: 21,
         name: 'test',
       });
-    });
-  });
-
-  describe('publish: createGitDirectory', () => {
-    const values = {
-      isOrg: true,
-      storePath: 'blam/test',
-      owner: 'lols',
-    };
-
-    const mockDir = '/tmp/test/dir';
-
-    mockGitlabClient.Projects.create.mockResolvedValue({
-      http_url_to_repo: 'mockclone',
-    } as { http_url_to_repo: string });
-
-    it('should call init on the repo with the directory', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(Repository.init).toHaveBeenCalledWith(mockDir, 0);
-    });
-
-    it('should call refresh index on the index and write the new files', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(mockRepo.refreshIndex).toHaveBeenCalled();
-    });
-
-    it('should call add all files and write', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(mockIndex.addAll).toHaveBeenCalled();
-      expect(mockIndex.write).toHaveBeenCalled();
-      expect(mockIndex.writeTree).toHaveBeenCalled();
-    });
-
-    it('should create a commit with on head with the right name and commiter', async () => {
-      const mockSignature = { mockSignature: 'bloblly' };
-      Signature.now.mockReturnValue(mockSignature);
-
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(Signature.now).toHaveBeenCalledTimes(2);
-      expect(Signature.now).toHaveBeenCalledWith(
-        'Scaffolder',
-        'scaffolder@backstage.io',
-      );
-
-      expect(mockRepo.createCommit).toHaveBeenCalledWith(
-        'HEAD',
-        mockSignature,
-        mockSignature,
-        'initial commit',
-        'mockoid',
-        [],
-      );
-    });
-
-    it('creates a remote with the repo and remote', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      expect(Remote.create).toHaveBeenCalledWith(
-        mockRepo,
-        'origin',
+      expect(pushToRemoteUserPass).toHaveBeenCalledWith(
+        '/tmp/test',
         'mockclone',
-      );
-    });
-
-    it('shoud push to the remote repo', async () => {
-      await publisher.publish({
-        values,
-        directory: mockDir,
-      });
-
-      const [remotes, { callbacks }] = mockRemote.push.mock
-        .calls[0] as NodeGit.PushOptions[];
-
-      expect(remotes).toEqual(['refs/heads/master:refs/heads/master']);
-
-      callbacks?.credentials?.();
-
-      expect(Cred.userpassPlaintextNew).toHaveBeenCalledWith(
         'oauth2',
         'fake-token',
       );
