+++ plugins/scaffolder-backend/src/scaffolder/stages/publish/gitlab.ts
@@ -16,10 +16,9 @@
 
 import { PublisherBase } from './types';
 import { Gitlab } from '@gitbeaker/core';
-
+import { pushToRemoteUserPass } from './helpers';
 import { JsonValue } from '@backstage/config';
 import { RequiredTemplateValues } from '../templater';
-import { Repository, Remote, Signature, Cred } from 'nodegit';
 
 export class GitlabPublisher implements PublisherBase {
   private readonly client: Gitlab;
@@ -38,7 +37,7 @@ export class GitlabPublisher implements PublisherBase {
     directory: string;
   }): Promise<{ remoteUrl: string }> {
     const remoteUrl = await this.createRemote(values);
-    await this.pushToRemote(directory, remoteUrl);
+    await pushToRemoteUserPass(directory, remoteUrl, 'oauth2', this.token);
 
     return { remoteUrl };
   }
@@ -62,29 +61,5 @@ export class GitlabPublisher implements PublisherBase {
     })) as { http_url_to_repo: string };
 
     return project?.http_url_to_repo;
-  }
-
-  private async pushToRemote(directory: string, remote: string): Promise<void> {
-    const repo = await Repository.init(directory, 0);
-    const index = await repo.refreshIndex();
-    await index.addAll();
-    await index.write();
-    const oid = await index.writeTree();
-    await repo.createCommit(
-      'HEAD',
-      Signature.now('Scaffolder', 'scaffolder@backstage.io'),
-      Signature.now('Scaffolder', 'scaffolder@backstage.io'),
-      'initial commit',
-      oid,
-      [],
-    );
-
-    const remoteRepo = await Remote.create(repo, 'origin', remote);
-
-    await remoteRepo.push(['refs/heads/master:refs/heads/master'], {
-      callbacks: {
-        credentials: () => Cred.userpassPlaintextNew('oauth2', this.token),
-      },
-    });
   }
 }
